Update to python3

+++ coffe/load_subcircuits.py
@@ -123,7 +123,7 @@ def local_routing_load_generate(spice_filename, num_on, num_partial, num_off):
     spice_file.write("******************************************************************************************\n")
     spice_file.write(".SUBCKT local_routing_wire_load n_in n_out n_gate n_gate_n n_vdd n_gnd n_vdd_local_mux_on\n")
     
-    num_total = num_on + num_partial + num_off
+    num_total = int(num_on + num_partial + num_off)
     interval_counter_partial = 0
     interval_counter_off = 0
     on_counter = 0
@@ -327,7 +327,7 @@ def generate_ble_outputs(spice_filename, num_local_out, num_gen_out):
     spice_file.write("******************************************************************************************\n")
     spice_file.write("* BLE outputs\n")
     spice_file.write("******************************************************************************************\n")
-    spice_file.write(".SUBCKT ble_outputs n_1_" + str((total_outputs + 1)/2+1) + " n_local_out n_general_out n_gate n_gate_n n_vdd n_gnd n_vdd_local_output_on n_vdd_general_output_on\n")
+    spice_file.write(".SUBCKT ble_outputs n_1_" + str(int((total_outputs + 1)/2)+1) + " n_local_out n_general_out n_gate n_gate_n n_vdd n_gnd n_vdd_local_output_on n_vdd_general_output_on\n")
     # Create the BLE output bar
     current_node = 2
     for i in range(num_local_out):
